test/programytest/storage/stores/sql/store/test_spelling.py
import unittest

from programytest.storage.asserts.store.assert_spelling import SpellingStoreAsserts

from programy.storage.stores.sql.store.spelling import SQLSpellingStore
from programy.storage.stores.sql.engine import SQLStorageEngine
from programy.storage.stores.sql.config import SQLStorageConfiguration

import programytest.storage.engines as Engines


class SQLSpellingStoreTests(SpellingStoreAsserts):

    @unittest.skipIf(Engines.sql is False, Engines.sql_disabled)
    def test_initialise(self):
        config = SQLStorageConfiguration()
        engine = SQLStorageEngine(config)
        engine.initialise()
        store = SQLSpellingStore(engine)
        self.assertEqual(store.storage_engine, engine)

    @unittest.skipIf(Engines.sql is False, Engines.sql_disabled)
    def test_upload_from_file(self):
        config = SQLStorageConfiguration()
        engine = SQLStorageEngine(config)
        engine.initialise()
        store = SQLSpellingStore(engine)

        self.assert_upload_from_file(store)
